requirements.txt
# Core Dependencies
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
imbalanced-learn==0.11.0
xgboost==2.0.0

# Kafka & Streaming
kafka-python==2.0.2
pyspark==3.4.1

# Database
pymongo==4.5.0

# Visualization & Dashboard
streamlit==1.27.0
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2

# Jupyter (optional for EDA)
jupyter==1.0.0
ipykernel==6.25.0
